+++ src/plugin.ts
@@ -274,6 +274,33 @@ Draw.loadPlugin(ui => {
     }
   }
 
+  // Visually disable edges connected to disabled nodes
+  // eslint-disable-next-line @typescript-eslint/unbound-method
+  const paintEdgeShape = mxConnector.prototype.paintEdgeShape;
+  mxConnector.prototype.paintEdgeShape = function(c: import('mxgraph').mxAbstractCanvas2D, pts: import('mxgraph').mxPoint[]) {
+    if (this.state && this.state.cell.source && this.state.cell.target) {
+      const source = AttributeRenderer.nodeAttributes(this.state.cell.source);
+      const target = AttributeRenderer.nodeAttributes(this.state.cell.target);
+
+      if (!source.getEnabledStatus() || !target.getEnabledStatus()) {
+        c.setAlpha(CellStyles.DISABLED_EDGE_ALPHA);
+      }
+    }
+    paintEdgeShape.apply(this, [c, pts]);
+  }
+
+  // Visually disable OR and AND nodes when disabled
+  // eslint-disable-next-line @typescript-eslint/unbound-method
+  const paintVertexShape = mxActor.prototype.paintVertexShape;
+  mxActor.prototype.paintVertexShape = function(c: import('mxgraph').mxAbstractCanvas2D, x: number, y: number, w: number, h: number) {
+    if (this.state
+        && CellStyles.isAttackgraphCell(this.state.cell)
+        && !AttributeRenderer.nodeAttributes(this.state.cell).getEnabledStatus()) {
+      c.setAlpha(CellStyles.DISABLED_CELL_ALPHA);
+    }
+    paintVertexShape.apply(this, [c, x, y, w, h]);
+  }
+
   /*
    * Highlight and mark edges connected to the selected node
    */